added 1d example
